% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnorm_ci_t.R
\name{abnorm_ci_t}
\alias{abnorm_ci_t}
\title{Confidence intervals for the abnormality of a test score using a modified t-test.}
\usage{
abnorm_ci_t(c, n, conf.level = 0.95)
}
\arguments{
\item{c}{An observation from a non-central t-distribution on N − 1 d.f.}

\item{n}{Size of the normative sample.}

\item{conf.level}{Confidence level (default is 0.95 for 95\%).}
}
\value{
A list containing the lower and upper bounds of the 95\% confidence interval for the abnormality of the test score.
}
\description{
Confidence intervals for the abnormality of a test score using a modified t-test.
}
\examples{
abnorm_ci_t(1.5, 30)
abnorm_ci_t(2, 10)
abnorm_ci_t(0.5, 100)

}
