% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aborm_j_battery.R
\name{aborm_j_battery}
\alias{aborm_j_battery}
\title{Calculate the percentage of the normative population with j or more abnormal test scores}
\usage{
aborm_j_battery(R, sims = 1e+05, threshold = -1.645)
}
\arguments{
\item{R}{A correlation matrix for the tests.}

\item{sims}{Number of simulations to perform. Default is 100000.}

\item{threshold}{Threshold for abnormal test scores. Default is -1.645 (corresponding to the 5th percentile in a standard normal distribution).}
}
\value{
A list containing:
\item{props}{Proportion of individuals with more than j-1 abnormal scores.}
\item{props.two.tail}{Proportion of individuals with more than j-1 abnormal scores, adjusted for a two-tailed test.}
}
\description{
This function calculates the percentage of the normative population that would exhibit
j or more abnormal test scores in a battery of tests. The method uses multivariate normal simulation.
}
\examples{
# Example usage with a 4x4 identity matrix as the correlation matrix
R <- diag(4)
aborm_j_battery(R)
}
