% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodpress.R
\name{bloodpress_mean}
\alias{bloodpress_mean}
\title{Calculate the mean blood pressure}
\usage{
bloodpress_mean(data, cols, na.rm = TRUE)
}
\arguments{
\item{data}{data frame}

\item{cols}{columns selected with tidy selectors}

\item{na.rm}{logical. Should missing values (including
NaN) be omitted from the calculations?}
}
\value{
numeric vector with mean
}
\description{
functions that given a data.frame
and columns selection in argument 'cols'
using tidy selectors, will calculate the mean
}
\examples{
  dt <- data.frame(
  BloodPress_Diastolic_1 = c(80,32,66,NA),
  BloodPress_Diastolic_2 = c(58,45,NA,99),
  BloodPress_Systolic_1 = c(40,NA,80,120),
  BloodPress_Systolic_2 = c(NA, 65,45,100)
  )
  
  bloodpress_mean(dt, dplyr::contains("Diastolic"))
  bloodpress_mean(dt, dplyr::contains("Systolic"))
}
\seealso{
Other blood pressure functions: \code{\link{bloodpress_map}}
}
\concept{blood pressure functions}
