% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi_calc}
\alias{bmi_calc}
\title{Calculate BMI}
\usage{
bmi_calc(height, weight)
}
\arguments{
\item{height}{height in centimeters}

\item{weight}{weight in kilograms}
}
\value{
numeric vector of BMI
}
\description{
function to calculate the BMI
with the standard formulae: 
\eqn{BMI = weight / (height/100)^2 }
}
\examples{
bmi_calc(176, 72)
}
\seealso{
Other bmi functions: \code{\link{bmi_calc2}}
}
\concept{bmi functions}
