% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-functions.R
\name{bmi_calc2}
\alias{bmi_calc2}
\title{Calculate BMI}
\usage{
bmi_calc2(height, weight, unit = list(height = "cm", weight = "kg"))
}
\arguments{
\item{height}{height}

\item{weight}{weight}

\item{unit}{list of weight and heigt with units for the measures}
}
\value{
numeric vector of BMI
}
\description{
function to calculate the BMI with the non-standard formulae: \eqn{BMI =
weight / (height/100)^(2.5) } which is suggested to provide a measurement that
it less dependent on height.
}
\examples{
bmi_calc2(176, 72)
}
\seealso{
Other bmi functions: 
\code{\link{bmi_calc}()}
}
\concept{bmi functions}
