% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{calc_bmi}
\alias{calc_bmi}
\title{Calculate BMI}
\usage{
calc_bmi(height, weight)
}
\arguments{
\item{height}{height in centimeters}

\item{weight}{weight in kilograms}
}
\value{
numeric vector of BMI
}
\description{
function to calculate the BMI
with the standard formulae: 
\eqn{BMI = weight / (height/100)^2 }
}
\examples{
calc_bmi(176, 72)
}
