% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{calc_map_mean}
\alias{calc_map_mean}
\title{Calculate mean arterial pressure}
\usage{
calc_map_mean(DATA)
}
\arguments{
\item{DATA}{data frame containing necessary columns}
}
\description{
Calculate mean arterial pressure
}
