% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{Calc_MapMean}
\alias{Calc_MapMean}
\title{Calculate mean arterial pressure}
\usage{
Calc_MapMean(DATA)
}
\arguments{
\item{DATA}{data frame containing necessary columns}
}
\description{
Calculate mean arterial pressure
}
