% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convert Celsius to Kelvin}
\usage{
celsius_to_kelvin(temp_C)
}
\arguments{
\item{temp_C}{A number. Do not use special characters or symbols.}
}
\value{
A temperature in degrees Kelvin as a function of that temperature in
degrees Celsius
}
\description{
Convert Celsius to Kelvin
}
\examples{
celsius_to_kelvin(0)
celsius_to_kelvin(32)
}
