% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_norms_single_pred.R
\name{cont_norms_single_pred}
\alias{cont_norms_single_pred}
\title{Generate Regression Norms for a Single Variable}
\usage{
cont_norms_single_pred(
  ctrl.x.mean,
  ctrl.x.sd,
  ctrl.y.mean,
  ctrl.y.sd,
  r,
  n,
  x,
  y,
  conf.level = 0.05,
  direction = "lower",
  dp = 4,
  x.name = "X",
  y.name = "Y"
)
}
\arguments{
\item{ctrl.x.mean}{Mean of the predictor variable in the control sample.}

\item{ctrl.x.sd}{Standard deviation of the predictor variable in the control sample.}

\item{ctrl.y.mean}{Mean of the criterion variable in the control sample.}

\item{ctrl.y.sd}{Standard deviation of the criterion variable in the control sample.}

\item{r}{Correlation coefficient between the predictor and criterion variables in the control sample.}

\item{n}{Sample size of the control group.}

\item{x}{Score of the individual on the predictor variable.}

\item{y}{Obtained score of the individual on the criterion variable.}

\item{conf.level}{Confidence level for the significance test. Default is 0.05.}

\item{direction}{Direction for the significance test. Options are "lower" or "upper". Default is "lower".}

\item{dp}{Number of decimal places.}

\item{x.name}{Name for the predictor variable. Default is "X".}

\item{y.name}{Name for the criterion variable. Default is "Y".}
}
\value{
A list containing calculated values and results.
}
\description{
This function generates regression norms and performs significance testing for an individual's score on a predictor variable. It calculates the regression equation, standard errors, and significance levels based on summary statistics from a control sample.
}
\examples{
cont_norms_single_pred(
  ctrl.x.mean = 63.8,
  ctrl.x.sd = 8.42,
  ctrl.y.mean = 41.3,
  ctrl.y.sd = 13.2,
  r = -0.58,
  n = 160,
  x = 26,
  y = 41,
  x.name = "Predictor",
  y.name = "Criterion"
)
}
