% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveRsion.R
\name{conveRsion}
\alias{conveRsion}
\title{Conversion between list of measures}
\usage{
conveRsion(x, from, to)
}
\arguments{
\item{infile}{An Integer for from argument}
}
\value{
An integer for to argument
}
\description{
function converts from to to unit for a given categorz.
It takes integer and returns integer.
}
