% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq_functions.R
\name{convert_rawIQ}
\alias{convert_rawIQ}
\title{Convert raw IQ measures to T or scaled}
\usage{
convert_rawIQ(data, type = "T", battery = "WASI", subtest = "Vocab",
  table = NULL, keep = F, ...)
}
\arguments{
\item{data}{data.frame with raw data}

\item{type}{string indicating type of conversion T [default] or scaled}

\item{battery}{from which test battery 'WASI'[default] or 'WPPSI'}

\item{subtest}{which subtest of battery}

\item{table}{Table with conversions, either directly as data.frame or path to file}

\item{keep}{unused option as of now}

\item{...}{extra options to rio::import for table import}
}
\description{
Convert raw IQ measures to T or scaled
}
