% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_standard.R
\name{convert_standard}
\alias{convert_standard}
\title{Convert Between Standardised Test Scores.}
\usage{
convert_standard(
  score,
  metric = "z",
  metric.new = "index",
  mean.new = NULL,
  sd.new = NULL,
  dp = 2
)
}
\arguments{
\item{score}{Numeric value (or vector) containing the standardised scores to
be converted.}

\item{metric}{Character string specifying the metric of the input scores.
("z", "t", "index", "scaled", "sten", "stanine", "sat", "percentile").
Default is "z".}

\item{metric.new}{Character string specifying the metric to convert input
scores to. Choose from "z", "t", "index", "scaled", "sten", "stanine",
"sat", "percentile". Default is "z".}

\item{mean.new}{Optional numeric value specifying the mean of the new metric.
If provided, this takes precedence over \code{metric.new}.}

\item{sd.new}{Optional numeric value specifying the standard deviation of the
new metric. If provided, this takes precedence over \code{metric.new}.}

\item{dp}{Integer specifying the number of decimal places to round the
converted scores to. Default is 2.}
}
\value{
A numeric value or vector of converted scores.
}
\description{
This function converts standardised scores between various metrics commonly
used in psychological and educational testing. The conversion supports
several score types including Z scores, T scores, index scores, scaled
scores, sten scores, stanine scores, SAT scores, and percentile ranks.
}
\details{
The function is flexible, allowing for custom mean and standard deviation
inputs for the new metric, making it suitable for a wide range of
standardised testing applications.

The conversion process uses a linear transformation based on the
properties of the normal distribution. Specifically, the conversion is
performed using the formula (as described in Crawford, 2004):

\deqn{
X_{\text{new}} = \frac{\sigma_{\text{new}}}{\sigma_{\text{old}}} \times (X_{\text{old}} - \bar{X}_{\text{old}}) + \bar{X}_{\text{new}}
}

Where:

\eqn{X_{\text{new}}} is the converted score,

\eqn{X_{\text{old}}} is the original score,

\eqn{\bar{X}_{\text{old}}} and \eqn{\sigma_{\text{old}}} are the mean and standard deviation of the original metric.

\eqn{\bar{X}_{\text{new}}} and \eqn{\sigma_{\text{new}}} are the mean and standard deviation of the new metric.

Below is a table summarizing the default mean and standard deviation for each metric:\tabular{lll}{
   Metric \tab Mean \tab Standard Deviation \cr
   Z-score \tab 0 \tab 1 \cr
   T-score \tab 50 \tab 10 \cr
   Index Score \tab 100 \tab 15 \cr
   Scaled Score \tab 10 \tab 3 \cr
   Sten Score \tab 5.5 \tab 2 \cr
   Stanine \tab 5 \tab 2 \cr
   SAT Score \tab 500 \tab 100 \cr
   Percentile \tab 50 \tab N/A (non-linear) \cr
}
}
\note{
Ensure that the \code{metric} and \code{metric.new} are correctly specified, as
incorrect metrics may lead to inappropriate score conversions. The function
assumes that the input scores are correctly standardised according to the
specified input metric.
}
\examples{
# Convert a Z score to a T score
convert_standard(score = 2, metric = "z", metric.new = "t", dp = 2)

# Convert a Z score to a Percentile Rank
convert_standard(score = 2, metric = "z", metric.new = "percentile", dp = 2)

# Convert a Z score to a custom metric with mean 60 and standard deviation 15
convert_standard(score = 2, metric = "z", mean.new = 60, sd.new = 15, dp = 2)

# Convert multiple scores from T scores to scaled scores
convert_standard(score = c(50, 55, 60), metric = "t", metric.new = "scaled", dp = 2)

}
\references{
\itemize{
\item Crawford, J. R. (2013). Quantitative aspects of neuropsychological assessment. In L. H. Goldstein & J. E. McNeil (Eds.), Clinical neuropsychology: A practical guide to assessment and management for clinicians (2nd ed., pp. 97-121). Wiley-Blackwell.
}
}
\seealso{
\code{\link[=convert_z]{convert_z()}}
}
