% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_z.R
\name{convert_z}
\alias{convert_z}
\title{Convert a z score to an alternative derived score.}
\usage{
convert_z(z, metric = "index", dp = 2)
}
\arguments{
\item{z}{numeric input (or vector) specifying the Z score for conversion.}

\item{metric}{Character string specifying the metric to convert input scores to. Choose from "z", "t", "index", "scaled", "sten", "stanine", "sat", or "percentile". Default is "index".}

\item{dp}{Integer specifying the number of decimal places to round the input scores to. Default is 2.}
}
\value{
a numeric value
}
\description{
This function takes a z score and converts it to one of a number of alternative
derived scores (e.g., T score, index score, scaled score, sten score, stanine score, percentile).
}
\examples{
convert_z(z = 2, metric = "index")
convert_z(z = 0.32, metric = "sten")
convert_z(z = 1, metric = "percentile")
}
