% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deficit_bayes.R
\name{deficit_bayes}
\alias{deficit_bayes}
\title{Bayesian Dissociation Single Case Analysis}
\usage{
deficit_bayes(
  score,
  ctrl.mean,
  ctrl.sd,
  ctrl.n,
  conf.level = 0.95,
  direction = "lower",
  tail = "one.tailed",
  dp = 2,
  sims = 10000,
  treshold = 0.1
)
}
\arguments{
\item{score}{Numeric value representing the score of the single case.}

\item{ctrl.mean}{Numeric value representing the mean of the control group.}

\item{ctrl.sd}{Numeric value representing the standard deviation of the
control group.}

\item{ctrl.n}{Integer value representing the sample size of the control
group.}

\item{conf.level}{Numeric value specifying the confidence level for the credible interval (default is 0.95 for 95\%).}

\item{direction}{Character. Specifies the direction of the hypothesis.
Options are "lower" (one-tailed), "higher" (one-tailed), or "two.tailed"
(default, two-tailed).}

\item{tail}{Character. Specifies whether the test is one-tailed or
two-tailed. Options are "one.tailed" and "two.tailed" (default)}

\item{dp}{Number of decimal places for rounding the results (default is 2).}

\item{sims}{Integer specifying the number of simulations to perform. Default is 10000.}

\item{treshold}{Numeric value for the abnormality threshold. Default is 0.1.}
}
\value{
A list of statistical input, parameters, and results. Key outputs
include:
\itemize{
\item t value: The t-value calculated for the test.
\item p value: The p-value for the test, indicating statistical significance.
\item effect-size (z-cc): The z-score (effect-size) corrected for the control group.
\item abnormality: The percentage of the population expected to score a more extreme score.
}
}
\description{
This function performs a Bayesian single-case analysis to assess the
abnormality of a given score compared to a control group. It generates
simulated distributions of scores based on the control group's mean,
standard deviation, and sample size, and provides Bayesian credible
intervals, p-values, and abnormality percentages.
}
\details{
NEEDS WRITING
}
\examples{
deficit_bayes(
  score = 90,
  ctrl.mean = 100,
  ctrl.sd = 15,
  ctrl.n = 30,
  conf.level = 0.95,
)

}
\references{
\itemize{
\item Crawford, J.R., & Garthwaite, P.H. (2007). Comparison of a single case to a control or normative sample in neuropsychology: Development of a Bayesian approach. \emph{Cognitive Neuropsychology, 24}(4), 343-372.
\item NEEDS WRITING.
\item NEEDS WRITING.
\item NEEDS WRITING.
\item NEEDS WRITING.
}
}
\seealso{
\itemize{
\item \code{\link[=deficit]{deficit()}}: Assessing For a frequentist single dissociation between a test score and a control sample.
\item \code{\link[=deficit_bayes]{deficit_bayes()}}: For a Bayesian approach to assessing for a dissociation between a single test score and a control sample for a single case.
\item \code{\link[=discrep]{discrep()}}: For assessing a dissociation between two test scores for a single case.
\item \code{\link[=abnorm_ci_t]{abnorm_ci_t()}}: For generating interval estimates for abnormality using the modified t test.
\item \code{\link[bayestestR:hdi]{bayestestR::hdi()}}
}
}
