% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrep_usdt.R
\name{discrep_usdt}
\alias{discrep_usdt}
\title{USDT}
\usage{
discrep_usdt(
  ctrl.mean.x,
  ctrl.sd.x,
  ctrl.mean.y,
  ctrl.sd.y,
  ctrl.r.xy,
  ctrl.n,
  score.x,
  score.y,
  direction = "lower",
  tail = "one.tailed",
  test.names = c("test X", "test Y"),
  conf.level = 0.95,
  dp = 2
)
}
\arguments{
\item{ctrl.mean.x}{Mean of the control group for the first test.}

\item{ctrl.sd.x}{Standard deviation of the control group for the first test.}

\item{ctrl.mean.y}{Mean of the control group for the second test.}

\item{ctrl.sd.y}{Standard deviation of the control group for the second test.}

\item{ctrl.r.xy}{Correlation between the two tests in the control group.}

\item{ctrl.n}{Integer value representing the sample size of the control
group.}

\item{score.x}{Test score of the individual case for the first test.}

\item{score.y}{Test score of the individual case for the second test.}

\item{direction}{Character. Specifies the direction of the hypothesis.
Options are "lower" (one-tailed), "higher" (one-tailed), or "two.tailed"
(default, two-tailed).}

\item{tail}{Character. Specifies whether the test is one-tailed or
two-tailed. Options are "one.tailed" and "two.tailed" (default)}

\item{test.names}{A vector of two strings representing the names of the tests
(default is c("X", "Y")).}

\item{conf.level}{Confidence level (default is 0.95 for 95\%).}

\item{dp}{Number of decimal places for rounding the results (default is 2).}
}
\value{
A list of statistical input, parameters, and results. Key outputs
include:
\itemize{
\item t value: The t-value calculated for the test.
\item p value: The p-value for the test, indicating statistical significance.
\item effect-size (z-cc): The z-score (effect-size) corrected for the control group.
\item abnormality: The percentage of the population expected to score a more extreme score.
}
}
\description{
description.
}
\details{
...
}
\examples{
discrep_usdt(100, 15, 110, 10, 0.5, 30, 130, 120, test.names = c("Fluency", "Sequencing"))
}
\references{
\itemize{
\item To complete
\item To complete
}
}
\seealso{
\itemize{
\item \code{\link[=dissoc_discrep]{dissoc_discrep()}}:
\item \code{\link[=dissoc_single]{dissoc_single()}}:
}
}
