% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissociation_double.R
\name{dissociation_double}
\alias{dissociation_double}
\title{Calculate abnormality of two test scores with confidence intervals using the modified t-test.}
\usage{
dissociation_double(
  ctrl.mean.x,
  ctrl.sd.x,
  ctrl.mean.y,
  ctrl.sd.y,
  ctrl.r.xy,
  ctrl.n,
  score.x,
  score.y,
  conf.level = 0.95,
  direction = "lower",
  dp = 4,
  test.names = c("X", "Y")
)
}
\arguments{
\item{ctrl.mean.x}{Mean of the control group for the first test.}

\item{ctrl.sd.x}{Standard deviation of the control group for the first test.}

\item{ctrl.mean.y}{Mean of the control group for the second test.}

\item{ctrl.sd.y}{Standard deviation of the control group for the second test.}

\item{ctrl.r.xy}{Correlation between the two tests in the control group.}

\item{ctrl.n}{Size of the control group.}

\item{score.x}{Test score of the individual for the first test.}

\item{score.y}{Test score of the individual for the second test.}

\item{conf.level}{Confidence level for the interval (default is 0.05).}

\item{direction}{Direction of the test, either "lower" or "higher" (default is "lower").}

\item{dp}{Number of decimal places for rounding the results (default is 2).}

\item{test.names}{A vector of two strings representing the names of the tests (default is c("X", "Y")).}
}
\value{
A list containing the z-scores for both tests, control group means and standard deviations, correlation between tests, control group size, scores for both tests, confidence level, direction of the test, t-value, p-value, two-tailed p-value, zdcc, zdcc confidence interval lower bound, zdcc confidence interval upper bound, abnormality percentage, abnormality confidence interval lower bound, and abnormality confidence interval upper bound.
}
\description{
Calculate abnormality of two test scores with confidence intervals using the modified t-test.
}
\examples{
dissociation_double(100, 15, 110, 10, 0.5, 30, 130, 120, test.names = c("Test A", "Test B"))
}
