% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissociation_single.R
\name{dissociation_single}
\alias{dissociation_single}
\title{Calculate abnormality of a test score with confidence intervals using the modified t-test.}
\usage{
dissociation_single(
  score,
  ctrl.mean,
  ctrl.sd,
  ctrl.n,
  conf.level = 0.05,
  direction = "lower",
  dp = 5
)
}
\arguments{
\item{score}{Test score of the individual.}

\item{ctrl.mean}{Mean of the control group.}

\item{ctrl.sd}{Standard deviation of the control group.}

\item{ctrl.n}{Size of the control group.}

\item{conf.level}{Confidence level for the interval (default is 0.05).}

\item{direction}{Direction of the test, either "lower" or "higher" (default is "lower").}

\item{dp}{Number of decimal places for rounding the results (default is 5).}
}
\value{
A list containing the score, control group mean, control group standard deviation, control group size, confidence level, direction of the test, number of decimal places, t-value, p-value, two-tailed p-value, zcc, zcc confidence interval lower bound, zcc confidence interval upper bound, abnormality percentage, abnormality confidence interval lower bound, and abnormality confidence interval upper bound.
}
\description{
Calculate abnormality of a test score with confidence intervals using the modified t-test.
}
\examples{
dissociation_single(130, 100, 15, 30)
}
