% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissociation_single_bayes.R
\name{dissociation_single_bayes}
\alias{dissociation_single_bayes}
\title{Calculate abnormality of a test score with Bayesian confidence intervals.}
\usage{
dissociation_single_bayes(
  score,
  ctrl.mean,
  ctrl.sd,
  ctrl.n,
  conf.level = 0.95,
  sims = 10000,
  treshold = 0.1,
  direction = "lower",
  dp = 5
)
}
\arguments{
\item{score}{Test score of the individual.}

\item{ctrl.mean}{Mean of the control group.}

\item{ctrl.sd}{Standard deviation of the control group.}

\item{ctrl.n}{Size of the control group.}

\item{conf.level}{Confidence level for the interval (default is 0.95).}

\item{sims}{Number of simulations for the Bayesian estimation (default is 10000).}

\item{treshold}{Threshold for determining abnormality (default is 0.1).}

\item{direction}{Direction of the test, either "lower" or "higher" (default is "lower").}

\item{dp}{Number of decimal places for rounding the results (default is 5).}
}
\value{
A list containing the score, control group mean, control group standard deviation, control group size, confidence level, number of simulations, threshold, direction of the test, degrees of freedom, control group variance, z-value, z-value confidence interval lower bound, z-value confidence interval upper bound, percentile, p-value, two-tailed p-value, abnormality percentage, abnormality confidence interval lower bound, and abnormality confidence interval upper bound.
}
\description{
Calculate abnormality of a test score with Bayesian confidence intervals.
}
\examples{
dissociation_single_bayes(130, 100, 15, 30)
}
