% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\value{
A temperature in degrees Celsius as a function of that temperature in
degrees Fahrenheit
}
\description{
Convert Fahrenheit to Celsius
}
\examples{
fahrenheit_to_celsius(212)
fahrenheit_to_celsius(32)
}
