% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Convert Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(temp_F)
}
\arguments{
\item{temp_F}{A number. Do not use special characters or symbols.}
}
\value{
A temperature in degrees Kelvin as a function of that temperature in
degrees Fahrenheit
}
\description{
Convert Fahrenheit to Kelvin
}
\examples{
fahrenheit_to_kelvin(2)
fahrenheit_to_kelvin(76)
}
