% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{gallon_to_ltr}
\alias{gallon_to_ltr}
\title{Provides gallon to liters conversion}
\usage{
gallon_to_ltr(gal)
}
\arguments{
\item{gal}{A value parameter.}
}
\value{
ltr A value.
}
\description{
Provides gallon to liters conversion
}
