% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_composite.R
\name{generate_composite}
\alias{generate_composite}
\title{Generate Composite Scores}
\usage{
generate_composite(
  r,
  r.between,
  mean = 10,
  sd = 3,
  comp.mean = 100,
  comp.sd = 15,
  dp = 4
)
}
\arguments{
\item{r}{A numeric vector of correlations for the included tests.}

\item{r.between}{A numeric vector of inter-correlations between the included tests.}

\item{mean}{A numeric value specifying the mean of the included tests. Default is 10.}

\item{sd}{A numeric value specifying the standard deviation of the included tests. Default is 3.}

\item{comp.mean}{A numeric value specifying the mean of the composite score. Default is 100.}

\item{comp.sd}{A numeric value specifying the standard deviation of the composite score. Default is 15.}

\item{dp}{An integer specifying the number of decimal places for rounding results. Default is 4.}
}
\value{
A list of class 'gen_comp' containing:
\item{k}{Number of subtests.}
\item{r}{Vector of correlations for the included tests.}
\item{r.between}{Vector of inter-correlations between the included tests.}
\item{composite_r}{Composite correlation.}
\item{original_mean}{Original mean of the composite score.}
\item{original_sd}{Original standard deviation of the composite score.}
\item{original_sem}{Original standard error of measurement.}
\item{original_semt}{Original standard error of the mean transformed.}
\item{transformed.mean}{Transformed mean of the composite score.}
\item{transformed.sd}{Transformed standard deviation of the composite score.}
\item{transformed.sem}{Transformed standard error of measurement.}
\item{transformed.semt}{Transformed standard error of the mean transformed.}
}
\description{
This function calculates composite scores based on subtest correlations and their inter-correlations.
}
\examples{
result <- generate_composite(r = c(.87, .94, .94),
                             r.between = c(0.74, 0.64, 0.73),
                             dp = 3)
print(result)

}
