% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_descriptor.R
\name{generate_descriptor}
\alias{generate_descriptor}
\title{Generate a descriptor for a derived score.}
\usage{
generate_descriptor(score, metric = "index", system = "aan")
}
\arguments{
\item{score}{Numeric value (or vector) containing the index scores to be converted to descriptors.}

\item{metric}{Character string specifying the metric of the input scores. ("z", "t", "index", "scaled", "sten", "stanine", "sat"). Default is "index".}

\item{system}{Character string specifying the descriptor system to use. Default is "aan".}
}
\value{
A character vector containing the descriptors corresponding to the input index scores.
}
\description{
This function converts index scores to descriptive labels based on predefined threshold values for various descriptor systems such as AAN, WISC, WAIS, NEPSY, Q.Simple, Groth-Marnat, and Heaton.
}
\details{
The function uses predefined threshold values and descriptor labels for each descriptor system to convert index scores into descriptive categories. It allows customization of the descriptor system and rounding of input scores.
}
\examples{
# Convert index scores to descriptors using the AAN descriptor system
generate_descriptor(score = c(85, 105, 120), metric = "index", system = "aan")

# Convert index scores to descriptors using the NEPSY descriptor system
generate_descriptor(score = c(75, 95, 110), metric = "index", system = "nepsy")
}
