% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_discreps_from_mean.R
\name{generate_discreps_from_mean}
\alias{generate_discreps_from_mean}
\title{Calculate score deviances and significance}
\usage{
generate_discreps_from_mean(
  x,
  R = NULL,
  sem,
  sd = 15,
  dp = 2,
  names = NULL,
  conf.level = 0.9,
  threshold = -1.645,
  apply_bonferroni = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{x}{A numeric vector of test scores.}

\item{R}{A numeric matrix representing the correlation matrix. If not provided, correlation-based calculations are skipped.}

\item{sem}{A numeric vector of standard error of measurement (SEM) values.}

\item{sd}{Standard deviation of the test scores. Default is 15.}

\item{dp}{Number of decimal places for rounding in the results. Default is 2.}

\item{names}{Optional character vector of names for the test scores. If not provided, default names will be used.}

\item{conf.level}{Confidence level for the confidence intervals. Default is 0.90.}

\item{threshold}{Threshold for abnormality detection. Default is -1.645.}

\item{apply_bonferroni}{Logical, whether to apply the Bonferroni correction. Default is FALSE.}

\item{alpha}{Significance level used in the Bonferroni correction. Default is 0.05.}
}
\value{
A list containing various statistics and results.
}
\description{
This function calculates the mean deviances of test scores from the mean,
evaluates their significance, and computes various related statistics.
}
\examples{
sem <- c(3.000000, 3.354102, 3.674235, 4.743416)
scores <- c(118, 107, 77, 68)
R <- matrix(c(1.00, 0.61, 0.64, 0.45,
              0.61, 1.00, 0.62, 0.52,
              0.64, 0.62, 1.00, 0.51,
              0.45, 0.52, 0.51, 1.00), nrow = 4, byrow = TRUE)
names <- c("Verbal Comprehension", "Perceptual Reasoning",
           "Working Memory", "Processing Speed")
generate_discreps_from_mean(scores, R = R, sem = sem, conf.level = 0.95, names = names)
}
