% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_percentile_norms.R
\name{generate_percentile_norms}
\alias{generate_percentile_norms}
\title{Calculate Percentile Norms}
\usage{
generate_percentile_norms(dat, method = "all", range = NULL)
}
\arguments{
\item{dat}{A numeric vector representing the dataset for which percentile norms are to be calculated.}

\item{method}{A character string specifying the method for calculating percentiles: "a" for method a, "b" for method b, "c" for method c, or "all" for all methods. Default is "all".}

\item{range}{An optional numeric vector specifying the range of values over which to calculate percentiles, or a range specified as a sequence (e.g., (1:12)). Default is the range from the minimum to the maximum value in the dataset.}
}
\value{
A data frame containing the following columns:
\describe{
\item{raw}{The raw values for which percentiles are calculated.}
\item{count}{The count of occurrences of each raw value in the dataset.}
\item{definition_a}{The percentile values calculated using method "a".}
\item{definition_b}{The percentile values calculated using method "b".}
\item{definition_c}{The percentile values calculated using method "c".}
}
}
\description{
This function calculates percentile norms for a given dataset.
}
\examples{
# Generate example data
set.seed(123)
data <- rnorm(100)

# Calculate percentile norms using all methods
generate_percentile_norms(data, method = "all")

# Calculate percentile norms using a specific method
generate_percentile_norms(data, method = "c")

# re-create Crawford (2009) table
crawford_2009 <- c(rep(0:12, c(0, 0, 0, 0, 0, 2, 4, 4, 14, 16, 20, 30, 10)))
generate_percentile_norms(crawford_2009)
}
