% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq_functions.R
\name{iq_raw2score}
\alias{iq_raw2score}
\title{Convert raw IQ scores to scaled or T}
\usage{
iq_raw2score(x, age, iq_table)
}
\arguments{
\item{x}{raw iq score}

\item{age}{age in decimals}

\item{iq_table}{table with converions}
}
\value{
numeric vector of converted IQ scores
}
\description{
Will convert raw IQ scores from subtests into T or scaled values according to
a conversion table
}
\examples{
\dontrun{
t <- iq_table(".iq_table_subtest.tsv", "Vocab", header=TRUE)
iq_raw2score(31, 22, t)
iq_raw2score(x = c(33, 34, NA, 34), age=c(15.5, 20, 20, NA))
}
}
\seealso{
Other iq-functions: \code{\link{iq_table}}
}
\concept{iq-functions}
