% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq_functions.R
\name{iq_t2iq}
\alias{iq_t2iq}
\title{Convert T-score to IQ}
\usage{
iq_t2iq(data, cols = NULL, iq_table = NULL)
}
\arguments{
\item{data}{data.frame}

\item{cols}{columns in the data frame with necessary
data}

\item{iq_table}{table with conversion, first column
being the score to convert from, second score to 
convert to}
}
\value{
numeric vector of IQ scores
}
\description{
This function converts iq T-scores
to IQ, using the conversion table provided.
The conversion table provided, and the columns
in the data selected through the 'cols' argument
must correspond regarding the test battery used,
and the number of subtests provided for the IQ.
If providing 2 WASI subtests scores, the conversion
table must be for the conversion of two subtests
to fullscale IQ. For WPPSI, you must select
columns with verbal and performance IQ scaled
to calculate the unscaled verbal and performane IQ.
For WPPSI fullscale IQ, apply the \code{iq_wppsi_fs}
function, using the two unscaled verbal and 
performance IQs.
}
\examples{
\dontrun{
##
} 
}
