% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq_functions.R
\name{iq_table}
\alias{iq_table}
\title{Import IQ conversion table}
\usage{
iq_table(table = NULL, subtest = NULL, ...)
}
\arguments{
\item{table}{path or data.frame with conversion data}

\item{subtest}{character vector indicating which subtest}

\item{...}{arguments to \code{rio::import}}
}
\value{
long tibble of the wanted conversion table
}
\description{
Import a punched version of the IQ
conversion table, for scaling raw
scores to norm or T-scores
}
\examples{
\dontrun{
conversion_table <- iq_table("tests/testthat/iq_table_subtest.tsv", header=TRUE)
iq_table(conversion_table, "vocabulary")
}
}
\seealso{
Other iq-functions: \code{\link{iq_raw2score}}
}
\concept{iq-functions}
