% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq-functions.R
\name{iq_wppsi_adjust}
\alias{iq_wppsi_adjust}
\title{Adjust WPPSI components to two subtest}
\usage{
iq_wppsi_adjust(scaled1, scaled2)
}
\arguments{
\item{scaled1, scaled2}{scaled score from subtest}
}
\value{
scaled verbal/performance iq
}
\description{
WPPSI requires 3 or more subtests for 
verbal and performance IQ. There is an
adjustment that may be made for it to 
approximate using two subtests. This function 
applies this adjustment.
}
\examples{
iq_wppsi_adjust(c(10, 14), c(14, 16))
}
