% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iq-functions.R
\name{iq_wppsi_fs}
\alias{iq_wppsi_fs}
\title{Calculate full scale IQ from WPPSI
verbal and performance IQ}
\usage{
iq_wppsi_fs(verbal_iq, performance_iq)
}
\arguments{
\item{verbal_iq}{unscaled verbal IQ}

\item{performance_iq}{unscaled performance IQ}
}
\value{
vector of full scale iq
}
\description{
Calculate full scale IQ from WPPSI
verbal and performance IQ
}
\examples{
iq_wppsi_fs(89, 96)
}
