% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdi.R
\name{mdi}
\alias{mdi}
\title{Calculate the Mahalanobis Distance Index (MDI)}
\usage{
mdi(R, x, norm.m, norm.sd, dp = 7)
}
\arguments{
\item{R}{A correlation matrix for the tests.}

\item{x}{A numeric vector of test scores.}

\item{norm.m}{A numeric vector of normative means.}

\item{norm.sd}{A numeric vector of normative SDs.}

\item{dp}{Number of decimal places for rounding in the results. Default is 7.}
}
\value{
A list containing:
\item{mdi}{The Mahalanobis Distance Index.}
\item{p_value}{The p-value associated with the MDI.}
\item{prev}{The percentage of the normative population expected to exhibit a more unusual profile.}
}
\description{
This function calculates the Mahalanobis Distance Index (MDI) for a given test score vector,
a correlation matrix, and the normative mean. It also computes the p-value and the percentage
of the normative population expected to exhibit a more unusual profile.
}
\examples{
# Example usage with a correlation matrix, test scores, and normative means
R <- diag(4) # Example correlation matrix
x <- c(118, 107, 77, 68)
norm.m <- c(100, 100, 100, 100)
norm.sd <- c(15, 15, 15, 15)
mdi_results <- mdi(R, x, norm.m = norm.m, norm.sd = norm.sd)
print(mdi_results)
}
