% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixingratio_to_percc.R
\name{mixingratio_to_percc}
\alias{mixingratio_to_percc}
\title{Convert Concentrations}
\usage{
mixingratio_to_percc(conc, temp = 298, pressure = 1, mixingratio = "ppb")
}
\arguments{
\item{conc}{numeric containing concentration value}

\item{temp}{tmeperature in K}

\item{pressure}{Pressure in atm}

\item{mixingratio}{Mixing ratio units i.e. ppth, ppm, ppb, ppt, ppq}
}
\value{
A numeric value in molecules cm-3
}
\description{
This function can have a mixing ratio input i.e. ppb, and output a concentration that in
per cubic centimeter. All concentrations should be stated. All temperatures should be in K
all pressures in atm.
}
