% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_Date.R
\name{new_Date}
\alias{new_Date}
\title{Load a Date}
\usage{
new_Date(x = double())
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A list of dates
}
\description{
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
