% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percc_to_mixingratio.R
\name{percc_to_mixingratio}
\alias{percc_to_mixingratio}
\title{Percc to Mixing Ratio}
\usage{
percc_to_mixingratio(conc, temp = 298, pressure = 1, mixingratio = "ppb")
}
\arguments{
\item{conc}{numeric containing concentration value}

\item{temp}{tmeperature in K}

\item{pressure}{Pressure in atm}

\item{mixingratio}{Mixing ratio units i.e. ppth, ppm, ppb, ppt, ppq}
}
\value{
A numeric value in molecules cm-3
}
\description{
This function can have a mixing ratio input i.e. ppb, and output a concentration that in
per cubic centimeter. All concentrations should be stated. All temperatures should be in K
all pressures in atm.
}
