% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{Calculate the percentile of a value in a dataset}
\usage{
percentile(data, x, method = "c")
}
\arguments{
\item{data}{A numeric vector representing the dataset.}

\item{x}{The value for which to calculate the percentile.}

\item{method}{The method to use for calculating the percentile. Options are
"a", "b", or "c". Defaults to "c".}
}
\value{
The percentile of the value in the dataset.
}
\description{
This function calculates the percentile of a given value in a dataset using
one of three methods: "a", "b", or "c".
}
\examples{
data <- c(1, 2, 3, 4, 5)
percentile(data, 3, method = "c")

}
