% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_vector.R
\name{percentile_vector}
\alias{percentile_vector}
\title{Calculate the percentiles of a vector of values}
\usage{
percentile_vector(data, method = "c")
}
\arguments{
\item{data}{A numeric vector representing the dataset.}

\item{method}{The method to use for calculating the percentiles.
Options are "a", "b", or "c". Defaults to "c".}
}
\value{
A numeric vector containing the percentiles of the input data.
}
\description{
This function calculates the percentiles of a vector of values using one of three methods: "a", "b", or "c".
}
\examples{
data <- c(1, 2, 3, 4, 5)
percentile_vector(data, method = "c")

}
