% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bell.R
\name{plot_bell}
\alias{plot_bell}
\title{Plot Bell Curve}
\usage{
plot_bell(
  input,
  metric = "z",
  score.label.text = "test",
  axis.label.metric = metric,
  descriptors = TRUE
)
}
\arguments{
\item{input}{A numeric value (or vector of numbers) representing the score of interest.}

\item{metric}{A character string specifying the metric of the input score: "percentile", "t", "index", "scaled", or "z" (default).}

\item{score.label.text}{A character string (or vector of character strings) for the label text of the input score(s).}

\item{axis.label.metric}{A character string specifying the preferred metric label for x-axis labels: "percentile", "t", "index", "scaled", or "z" (default).}

\item{descriptors}{A logical indicating whether to include AAN descriptor labels (default is TRUE).}
}
\value{
A ggplot2 object.
}
\description{
Plot Bell Curve
}
\examples{
plot_bell(input = 2, score.label.text = "Test Score", axis.label.metric = "z")
}
