% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reliable_change.R
\name{plot_reliable_change}
\alias{plot_reliable_change}
\title{Plot Stability}
\usage{
plot_reliable_change(
  data,
  metric = "z",
  group,
  test,
  t1.score,
  t2.expected,
  t2.score,
  t2.ci.lb,
  t2.ci.ub,
  axis.label.metric = metric,
  abbreviations = TRUE
)
}
\arguments{
\item{data}{The data frame containing the input data.}

\item{metric}{The metric to be used ("z" by default).}

\item{group}{The grouping variable in the data.}

\item{test}{The test variable in the data.}

\item{t1.score}{The score at time point 1.}

\item{t2.expected}{The expected score at time point 2.}

\item{t2.score}{The actual score at time point 2.}

\item{t2.ci.lb}{The lower bound of the confidence interval for time point 2.}

\item{t2.ci.ub}{The upper bound of the confidence interval for time point 2.}

\item{axis.label.metric}{The metric used for axis labeling (default is "z").}

\item{abbreviations}{Logical, whether to abbreviate test names (default is TRUE).}
}
\value{
A ggplot object representing the stability plot.
}
\description{
Plot Stability
}
