% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Custom ggplot2 theme}
\usage{
plot_theme()
}
\value{
A ggplot2 theme object.
}
\description{
This function provides a custom ggplot2 theme for consistent styling across plots.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  plot_theme()

}
