% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Calculate Prevalence}
\usage{
prevalence(z, method = "proportion", two.tailed = FALSE, dp = 5)
}
\arguments{
\item{z}{A numeric value representing the z-score.}

\item{method}{A character string specifying the method to calculate prevalence.
It must be one of "percentage", "proportion", or "ratio" (default: "ratio").}

\item{two.tailed}{Logical indicating whether to consider both tails of the distribution
when calculating prevalence (default: FALSE).}

\item{dp}{Integer specifying the number of decimal places for the result (default: 5).}
}
\value{
The prevalence calculated based on the specified method.
}
\description{
This function calculates the prevalence based on a given z-score.
}
\examples{
prevalence(z = -1, method = "proportion", two.tailed = TRUE, dp = 5)
prevalence(z = -1, method = "proportion", two.tailed = FALSE, dp = 5)
}
