% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_intervals_t.R
\name{prevalence_intervals_t}
\alias{prevalence_intervals_t}
\title{Confidence intervals for the abnormality of a test score using a modified t-test.}
\usage{
prevalence_intervals_t(c, n)
}
\arguments{
\item{c}{An observation from a non-central t-distribution on N − 1 d.f.}

\item{n}{Size of the normative sample.}
}
\value{
A list containing the lower and upper bounds of the 95\% confidence interval for the abnormality of the test score.
}
\description{
Confidence intervals for the abnormality of a test score using a modified t-test.
}
\examples{
prevalence_intervals_t(1.5, 30)
prevalence_intervals_t(2, 10)
prevalence_intervals_t(0.5, 100)
}
