% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_composite.R
\name{print.gen_comp}
\alias{print.gen_comp}
\title{Custom Print Method for 'gen_comp' Class}
\usage{
\method{print}{gen_comp}(x, ...)
}
\arguments{
\item{x}{An object of class 'gen_comp'.}

\item{...}{Additional arguments (not used).}
}
\description{
This method formats and prints the output for objects of class 'gen_comp'.
}
