% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdi.R
\name{print.mdi}
\alias{print.mdi}
\title{Custom Print Method for MDI}
\usage{
\method{print}{mdi}(x, ...)
}
\arguments{
\item{x}{An object containing the MDI results, which includes the
chi-square value, degrees of freedom (df), and p-value.}

\item{...}{Additional arguments passed to other methods (currently not used).}
}
\value{
The function prints the formatted MDI result but does not return a value.
}
\description{
This function provides a custom print method for displaying the results
of the Mahalanobis Distance Index (MDI) calculation in a formatted way.
}
