% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_discreps_from_mean.R
\name{print.mean_devs}
\alias{print.mean_devs}
\title{Print Method for Mean Deviances}
\usage{
\method{print}{mean_devs}(x, ...)
}
\arguments{
\item{x}{An object of class 'mean_devs'.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Custom print method for objects of class 'mean_devs'. This function
formats and prints the calculated deviances, confidence intervals,
p-values, and other related statistics using knitr::kable for better formatting.
}
