% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliable_change.R
\name{reliable_change}
\alias{reliable_change}
\title{Calculate reliable_change using various methods}
\usage{
reliable_change(
  method = "crawford",
  t1.score,
  norm.t1.mean,
  norm.t1.sd,
  norm.t2.mean,
  norm.t2.sd,
  norm.r,
  norm.n,
  conf.level = 0.9
)
}
\arguments{
\item{method}{The method used for reliable_change calculation. Default is "crawford".}

\item{t1.score}{The score at time 1.}

\item{norm.t1.mean}{Mean of the normative data at time 1.}

\item{norm.t1.sd}{Standard deviation of the normative data at time 1.}

\item{norm.t2.mean}{Mean of the normative data at time 2.}

\item{norm.t2.sd}{Standard deviation of the normative data at time 2.}

\item{norm.r}{Pearson correlation coefficient between time 1 and time 2 scores.}

\item{norm.n}{Sample size of the normative data.}

\item{conf.level}{Confidence level for the confidence interval calculation. Default is 0.9.}
}
\value{
A list containing the expected score at time 2, standard error of the mean,
lower and upper bounds of the confidence interval.
}
\description{
This function calculates reliable_change using different methods such as Crawford,
Jacobson, Speer, and others.
}
\examples{
reliable_change(method = "jacobson", 100, 103, 10, 109, 8, 0.76, 100, .95)

}
