% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_to_AH.R
\name{RH_to_AH}
\alias{RH_to_AH}
\title{AHtoRH}
\usage{
RH_to_AH(RH, temp = 25)
}
\arguments{
\item{RH}{#Numeric containing percentage Relative Humidity}

\item{temp}{#numeric containing temperature in Celcius}
}
\value{
A numeric value in g/m3
}
\description{
This function converts RHtoAH given a specfic temperature, pressure and RH
}
