% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_error.R
\name{se_mean}
\alias{se_mean}
\title{Standard Error of the Mean}
\usage{
se_mean(dat = NULL, sd = NULL, n = NULL)
}
\arguments{
\item{dat}{(numeric) A numeric vector from which to calculate standard
deviation and sample size. If provided, sd and n will be ignored.}

\item{sd}{(numeric) The standard deviation of the data. Ignored if dat is
provided.}

\item{n}{(numeric) The sample size. Ignored if dat is provided.}
}
\value{
The standard error of the mean.
}
\description{
Calculate the standard error of the mean given the standard deviation
and sample size, or calculate from data.
}
\examples{
# If dat is provided, calculate sd and n from dat
se_mean(c(1, 2, 3, 4, 5))

# If sd and n are provided
se_mean(sd = 3, n = 30)

# If dat is not provided but sd and n are provided
se_mean(sd = 3, n = 30)

}
