% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_error.R
\name{se_measurement}
\alias{se_measurement}
\title{Standard Error of Measurement (SEM)}
\usage{
se_measurement(dat = NULL, sd = NULL, r = NULL)
}
\arguments{
\item{dat}{(numeric) A numeric vector of measurements from which to calculate
the standard deviation. Ignored if r and SD are provided.}

\item{sd}{(numeric) The standard deviation of the measurements. If dat is
provided, sd will be calculated from dat.}

\item{r}{(numeric) The reliability of the test.}
}
\value{
The standard error of measurement (SEM).
}
\description{
Calculate the standard error of measurement (SEM) given the reliability and
standard deviation of scores
}
\details{
The Standard Error of Measurement is the standard deviation of an infinite
number of hypothetical scores around the true score. The SEM is calculated as
the standard deviation (SD) multiplied by the square root of 1 minus the test
retest reliability (r) of the measurement tool.
}
\examples{
# If dat and r are provided, calculate sd from dat and then SEM
se_measurement(dat = c(1, 2, 3, 4, 5), r = 0.4)

# If sd and r are provided, calculate SEM
se_measurement(sd = 3, r = 0.4)

}
