% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_error.R
\name{se_n1}
\alias{se_n1}
\title{Standard Error for n1}
\usage{
se_n1(sd.1, sd.2, r, n, t1.score, t2.score, norm.t1.mean)
}
\arguments{
\item{sd.1}{Standard deviation of the first variable.}

\item{sd.2}{Standard deviation of the second variable.}

\item{r}{Reliability coefficient (reliability of the test).}

\item{n}{Sample size.}

\item{t1.score}{Score of the first test.}

\item{t2.score}{Score of the second test.}

\item{norm.t1.mean}{Mean of the norm group for the first test.}
}
\value{
The standard error for sample size n=1.
}
\description{
This function calculates the standard error of the predicted score with
Crawford adjustment.
}
\examples{
# Example usage:
se_n1(sd.1 = 3, sd.2 = 4, r = 0.8, n = 100, t1.score = 75, norm.t1.mean = 70)

}
