% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_bonferroni.R
\name{sequential_bonferroni}
\alias{sequential_bonferroni}
\title{Sequential Bonferroni Correction}
\usage{
sequential_bonferroni(p.vals, alpha = 0.05)
}
\arguments{
\item{p.vals}{A numeric vector of p-values to be adjusted.}

\item{alpha}{A numeric value representing the significance level (default is 0.05).}
}
\value{
A list with the following components:
\item{sig}{A logical vector indicating which p-values are significant after correction.}
\item{adj.p.vals}{A numeric vector of adjusted p-values.}
}
\description{
Applies the Sequential Bonferroni correction to a vector of p-values.
}
\details{
This function applies the Sequential Bonferroni correction to control the
family-wise error rate (FWER) when performing multiple hypothesis tests. It adjusts
the significance threshold for each p-value based on its rank in the ordered list of p-values.
}
\examples{
# Example p-values
p.vals <- c(0.001, 0.04, 0.003, 0.08)

# Apply the sequential Bonferroni correction
results <- sequential_bonferroni(p.vals)

# Print significant results
results$sig

# Print adjusted p-values
results$adj.p.vals

}
