% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{stability_reg_est}
\alias{stability_reg_est}
\title{Regression Based Estimate for Stability}
\usage{
stability_reg_est(b, score, c)
}
\arguments{
\item{b}{A numeric value representing the beta}

\item{score}{A numeric value representing the score.}

\item{c}{A numeric value representing the constant.}
}
\value{
Returns the estimate of the time 2 score.
}
\description{
This function estimates stability regularization using the provided
parameters.
}
\examples{
stability_reg_est(0.5, 10, 3)

}
