% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_error.R
\name{stability_regression_estimate}
\alias{stability_regression_estimate}
\title{Regression Based Estimate for Stability}
\usage{
stability_regression_estimate(b, score, c)
}
\arguments{
\item{b}{A numeric value representing the beta}

\item{score}{A numeric value representing the score.}

\item{c}{A numeric value representing the constant.}
}
\value{
Returns the estimate of the time 2 score.
}
\description{
This function estimates stability regularization using the provided
parameters.
}
\examples{
stability_regression_estimate(0.5, 10, 3)

}
