% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodpress-functions.R
\name{bloodpress_map}
\alias{bloodpress_map}
\title{Calculate mean arterial pressure}
\usage{
bloodpress_map(diastolic, systolic)
}
\arguments{
\item{diastolic}{diastolic blood pressure}

\item{systolic}{systolic blood pressure}
}
\value{
numeric vector of mean arterial pressure
}
\description{
Calculates the mean arterial pressure
based on diastolic and systolic blood pressure.
\eqn{MAP = (diastolic * 2)  + systolic / 3 }
}
\examples{
bloodpress_map(69, 40)
}
\seealso{
Other blood pressure functions: 
\code{\link{bloodpress_mean}()}
}
\concept{blood pressure functions}
