% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveRsion.R
\docType{data}
\name{.listOfMeasures}
\alias{.listOfMeasures}
\title{Conversion between list of measures}
\format{
An object of class \code{data.frame} with 8 rows and 3 columns.
}
\usage{
.listOfMeasures
}
\arguments{
\item{infile}{An Integer for from argument}
}
\value{
An integer for to argument
}
\description{
function converts from to to unit for a given categorz.
It takes integer and returns integer.
}
\keyword{datasets}
