% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_standard.R
\name{convert_standard}
\alias{convert_standard}
\title{Convert a derived score to an alternative derived score.}
\usage{
convert_standard(score, metric = "index", metric.new = "index", dp = 2)
}
\arguments{
\item{score}{Numeric value (or vector) containing the index scores to be converted to descriptors.}

\item{metric}{Character string specifying the metric of the input scores. ("z", "t", "index", "scaled", "sten", "stanine", "sat"). Default is "index"}

\item{metric.new}{Character string specifying the metric to convert input scores to. Choose from "z", "t", "index", "scaled", "sten", "stanine", "sat", or "percentile". Default is "index".}

\item{dp}{Integer specifying the number of decimal places to round the input scores to. Default is 2.}
}
\value{
a numeric value
}
\description{
This function takes a derived score and converts it to one of a number of alternative
derived scores (Z score, T score, index score, scaled score, sten score, stanine score).
}
\examples{
convert_standard(score = 2, metric = "z", metric.new = "t", dp = 2)
convert_standard(score = 2, metric = "z", metric.new = "percentile", dp = 2)
}
