% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Convert Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(fahr)
}
\arguments{
\item{infile}{An Integer for Fahrenheit}
}
\value{
An integer for Kelvin
}
\description{
function converts Fahrenheit temperature to Kelvin temperature.
It takes integer and returns integer.
}
