% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliable_change.R
\name{reliable_change}
\alias{reliable_change}
\title{Discrepncy Between Predicted and Observed Time 2 Test Scores (Reliable
Change).}
\arguments{
\item{t1.score}{Scores of the first test.}

\item{t2.score}{Scores of the second test.}

\item{norm.t1.mean}{Mean of the normative sample at time 1.}

\item{norm.t1.sd}{Standard deviation of the normative sample at time 1.}

\item{norm.t2.mean}{Mean of the normative sample at time 2.}

\item{norm.t2.sd}{Standard deviation of the normative sample at time 2.}

\item{norm.r}{Reliability of the normative sample.}

\item{norm.n}{Sample size of the normative sample.}

\item{conf.level}{Confidence level for stability assessment.}

\item{methods}{A character vector specifying which methods to calculate.
Default is "all" methods.}
}
\description{
This function compares stability methods given the scores of two tests.
}
\examples{
# Example data
t1.score <- 88
norm.t1.mean <- 113
norm.t1.sd <- 7.2
norm.t2.mean <- 118.7
norm.t2.sd <- 8.3
norm.r <- 0.93
norm.n <- 821
t2.score <- 103
ci <- 0.95

# Using all methods
result_all <- reliable_change(t1.score, t2.score,
norm.t1.mean, norm.t1.sd, norm.t2.mean, norm.t2.sd, norm.r, norm.n,
ci)
print(result_all)

# Using selected methods
result_selected <- reliable_change(t1.score, t2.score,
norm.t1.mean, norm.t1.sd,norm.t2.mean, norm.t2.sd, norm.r,
norm.n, ci, methods = c("crawford", "maassen"))
print(result_selected)
}
